backend/requirements.txt
python-dotenv
langchain
fastapi[all]
octoai-sdk
nltk
pydantic
gunicorn
supabase
requests